requirements.txt
streamlit>=1.30
pandas
plotly
Pillow
